% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-magic-quote.R
\name{str_replace_magic_quotes}
\alias{str_replace_magic_quotes}
\title{Replace magic quotes with regular quotes}
\usage{
str_replace_magic_quotes(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector with quotes replaced
}
\description{
Replace magic quotes with regular quotes
}
\examples{
bad_text <- '“happens”'

str_replace_magic_quotes(bad_text)
}
